requirements.txt
Django
django-cors-headers
django-dotenv
djangorestframework
djangorestframework-simplejwt
PyJWT